Make log_group work with newer golang and libs
This switches the termui library to the latest, and removes the custom widget I had previously as the built in list widget comes close enough.

I also moved everything into the root of the repo, and removed log_multiline, as it's simple to implement in awk or similar. This also allows us to have a single go module for log_group, which I've now done. The repo will be renamed to log_group also.